Metadata on Ref properties
'$ref's needs to be entirely standalone because they are "true" JSON
reference objects. This means that in order to add *any* metadata to
them, you must wrap them in a `oneOf` key with one element.

Seriously.

This patch says "screw it, I don't wanna have to detect *when* things
happen always", and makes all refs wrapped in a oneOf key.