Fix #27 by adding `#merge`
This adds a new item to blueprint definitions, `#merge`, which merges in
a single level of output object. This hopefully works out pretty well.

I could see the use for a `deep_merge` that also inherits views, maybe?
I dunno.

Note: the way this is currently designed *will not* work with mutual
recursion as well, which might be problematic?