Warn on weird inheritance
Currently you might be tempted to try and define an array inline by
passing `SoberSwag::Types::Array` as the second arguemnt of `attribute`
or `attribute?`, then adding stuff underneath in the block. However,
this *does not work*, and also does not error.

This commit adds on functionality so this behavior results in an error
instead. More specifically, since the second argument of `attribute` is
used as a superclass when you pass a block, we yell at you if we don't
have some subclass of `SoberSwag::InputObject` for that.

Fixes #47