[Reporting] Initial Commit
Includes a lot of code for trying to do "reporting" inputs (which have
nicer errors when you pass invalid params) and outputs (which can
self-validate schema issues, whcih is nice).

This is really just an experiment at this point, but I am continuing to
work on it because it seems to have potential!